% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AATtools.R
\name{aat_bootstrap}
\alias{aat_bootstrap}
\title{Compute bootstrapped approach-bias scores}
\usage{
aat_bootstrap(ds, subjvar, pullvar, targetvar, rtvar, iters, plot = T,
  algorithm = c(aat_doublemeandiff, aat_doublemediandiff, aat_dscore,
  aat_multilevelscore), trialdropfunc = c(prune_nothing,
  trial_prune_3SD), errortrialfunc = c(prune_nothing,
  error_replace_blockmeanplus), ...)
}
\arguments{
\item{ds}{a longformat data.frame}

\item{subjvar}{Quoted name of the participant identifier column}

\item{pullvar}{Quoted name of the column indicating pull trials. 
Pull trials should either be represented by 1, or by the second level of a factor.}

\item{targetvar}{Name of the column indicating trials featuring the target stimulus. 
Target stimuli should either be represented by 1, or by the second level of a factor.}

\item{rtvar}{Name of the reaction time column.}

\item{iters}{Total number of desired iterations. At least 200 are required to get confidence intervals that make sense.}

\item{plot}{Plot the bias scores and their confidence intervals after computation is complete. This gives a good overview of the data.}

\item{algorithm}{Function (without brackets or quotes) to be used to compute AAT scores. See \link{aat_doublemeandiff} for a list of usable algorithms.}

\item{trialdropfunc}{Function (without brackets or quotes) to be used to exclude outlying trials in each half. 
\code{prune_nothing} excludes no trials, while \code{trial_prune_3SD} excludes trials deviating more than 3SD from the mean per participant.}

\item{errortrialfunc}{Function (without brackets or quotes) to apply to an error trial. 
\code{error_replace_blockmeanplus} replaces error trial reaction times with the block mean plus an arbitrary extra amount of time.
If used, the following additional arguments are required:
\itemize{
\item \code{blockvar} - Quoted name of the block variable
\item \code{errorvar} - Quoted name of the error variable, where errors are 1 or TRUE and correct trials are 0 or FALSE
\item \code{errorbonus} - Amount to add to the reaction time of error trials. Default is 0.6 (recommended by \code{Greenwald, Nosek, & Banaji, 2003})
}}

\item{...}{Other arguments, to be passed on to the algorithm functions (see \code{algorithm} above)}
}
\value{
A list, containing bootstrapped bias scores, a data frame with bootstrapped 95% confidence intervals, 
the number of iterations, and a matrix of bias scores for each iteration.
}
\description{
Compute bootstrapped approach-bias scores with confidence intervals.
}
\author{
Sercan Kahveci
}
