% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AATtools.R
\name{aat_doublemeandiff}
\alias{aat_doublemeandiff}
\alias{aat_doublemediandiff}
\alias{aat_dscore}
\alias{aat_multilevelscore}
\title{AAT score computation algorithms}
\usage{
aat_doublemeandiff(ds, subjvar, pullvar, targetvar, rtvar, ...)

aat_doublemediandiff(ds, subjvar, pullvar, targetvar, rtvar, ...)

aat_dscore(ds, subjvar, pullvar, targetvar, rtvar, ...)

aat_multilevelscore(ds, subjvar, formula, aatterm, ...)
}
\arguments{
\item{ds}{A long-format data.frame}

\item{subjvar}{Column name of the participant identifier variable}

\item{pullvar}{Column name of the movement variable (0: avoid; 1: approach)}

\item{targetvar}{Column name of the stimulus category variable (0: control stimulus; 1: target stimulus)}

\item{rtvar}{Column name of the reaction time variable}

\item{...}{Other arguments passed on by functions (ignored)}

\item{formula}{A character string containing a formula to fit to the data and derive multilevel scores from}

\item{aatterm}{The random term, grouped under the subject variable, which represents the approach bias. Usually this is the interaction of the pull and target terms.}
}
\value{
A data.frame containing participant number and computed AAT score.
}
\description{
\itemize{
\item \code{aat_doublemeandiff} computes a mean-based double-difference score: 

\code{(mean(push_target) - mean(pull_target)) - (mean(push_control) - mean(pull_control))}
\item \code{aat_doublemediandiff} computes a median-based double-difference score: 

\code{(median(push_target) - median(pull_target)) - (median(push_control) - median(pull_control))}
\item \code{aat_dscore} computes D-scores (see Greenwald, Nosek, and Banaji, 2003): 

\code{((mean(push_target) - mean(pull_target)) - (mean(push_control) - mean(pull_control))) / sd(participant_reaction_times)}
\item \code{aat_multilevelscore} fits a multilevel model using lme4 and extracts a random effect serving as AAT score. When using this function, additional arguments must be provided: 
\itemize{
\item \code{formula} - a quoted formula to fit to the data;
\item \code{aatterm} the quoted random effect within the subject variable that indicates the approach bias; this is usually the interaction of the pull and target terms.
}
}
}
