% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{allocate}
\alias{allocate}
\title{Quota-fair Sampler}
\usage{
allocate(x, size, prob = 1)
}
\arguments{
\item{x}{A vector to sample from.}

\item{size}{The desired length of the output vector.}

\item{prob}{(Optional) With what proportion should each value
in \code{x} occur in the output vector?}
}
\value{
A vector of length \code{size} with elements fairly sampled from \code{x}.
}
\description{
This samples \code{size} times from \code{x} such that the proportions given by
\code{prob} are met \emph{exactly}. When they cannot be met exactly, sampling occurs as
fairly as possible; in that case, only the observations that cannot be allocated fairly
are randomly sampled with probabilities such that many repetitions of the sampling
algorithm would yield the desired proportions. If no proportions are given,
all observations are sampled equally often.
}
\examples{
allocate(1:2,size=50)

test <- replicate(10000,allocate(1:5,size=23,prob=c(.1,.3,.2,.1,.3)))
table(test) / length(test)

}
\author{
Sercan Kahveci
}
