% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{allpairs}
\alias{allpairs}
\title{Generate unique pairs or N-tuplets}
\usage{
allpairs(nval, ntuplet = 2, incl.self = FALSE)
}
\arguments{
\item{nval}{Number of values to arrange into unique tuplets.}

\item{ntuplet}{N-tuplets to arrange the values uniquely into.}

\item{incl.self}{Determines whether a value can be paired with itself.}
}
\value{
A matrix where each row is a unique N-tuplet.
}
\description{
Generate unique pairs or N-tuplets
}
\examples{
allpairs(nval=5,ntuplet=3)

}
\author{
Sercan Kahveci
}
