% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4tools.R
\name{AnovaTable}
\alias{AnovaTable}
\alias{print.AnovaTable}
\title{Compare multilevel models}
\usage{
AnovaTable(..., serial = FALSE, suppress = c("AIC", "deviance", "logLik"))

\method{print}{AnovaTable}(x, ...)
}
\arguments{
\item{...}{lme4 model objects to be compared.}

\item{serial}{If \code{TRUE}, models are compared serially; 
if false, all models will be compared to the first.}

\item{suppress}{Character vector of column names to suppress in printed output.}

\item{x}{an \code{AnovaTable} object}
}
\value{
A data.frame containing model fit metrics such as AIC, BIC, 
marginal R-squared (the effect size of fixed effects only),
conditional R-squared (the effect size of all model terms), loglikelihood, 
deviance, and a likelihood ratio test.
}
\description{
Compare multilevel models
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(AnovaTable)}: Print generic for anova tables.

}}
\examples{
library(lmerTest)
data("sleepstudy", package="lme4")
m1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
m2 <- lmer(Reaction ~ Days + (1 | Subject), sleepstudy)
AnovaTable(m2,m1)

}
