% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skPlotting.R
\name{AutocorPlot}
\alias{AutocorPlot}
\title{Per-subject Autocorrelation Plotting}
\usage{
AutocorPlot(ds, ppvar, rtvar, scope = 64)
}
\arguments{
\item{ds}{a dataset}

\item{ppvar}{name of the variable indicating participant ID}

\item{rtvar}{name of the variable indicating reaction time}

\item{scope}{numeric, the maximum lag at which to compute autocorrelation.}
}
\description{
Per-subject Autocorrelation Plotting
}
\examples{
AutocorPlot(ds=ToothGrowth,ppvar="supp",rtvar="len",scope=10)
}
