% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faststats.R
\name{bonferroniHolm}
\alias{bonferroniHolm}
\alias{benjaminiHochberg}
\alias{benjaminiYekutieli}
\title{Adjusted p-values for false-discovery rates and familywise error rates}
\usage{
bonferroniHolm(x)

benjaminiHochberg(x)

benjaminiYekutieli(x)
}
\arguments{
\item{x}{A vector of p-values.}
}
\value{
A vector of p-values adjusted in accordance with the chosen method.
}
\description{
Adjusted p-values for false-discovery rates and familywise error rates
}
\examples{
oldp<-c(0.01,0.04,0.02,.1,0.035)
bonferroniHolm(oldp)
benjaminiHochberg(oldp)
benjaminiYekutieli(oldp)

}
\author{
Sercan Kahveci
}
