% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{carryforward}
\alias{carryforward}
\title{Carry non-NA values forward into NA values}
\usage{
carryforward(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
\code{x} with all \code{NA} values replaced with their 
last preceding non-\code{NA} value.
}
\description{
This replaces every \code{NA} value with their last preceding non-\code{NA} value.
}
\examples{
carryforward(c(NA,1,NA,NA,2))

}
\author{
Sercan Kahveci
}
