% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic.R
\name{Case-insensitive-operators}
\alias{Case-insensitive-operators}
\alias{\%cin\%}
\alias{\%ncin\%}
\alias{\%cis\%}
\alias{\%ncis\%}
\title{Case-insensitive inline operators}
\usage{
x \%cin\% y

x \%ncin\% y

x \%cis\% y

x \%ncis\% y
}
\arguments{
\item{x, y}{Values to be compared.}
}
\value{
A logical vector.
}
\description{
These operators convert both left- and right-hand side to lowercase 
before comparing.
}
\section{Functions}{
\itemize{
\item \code{x \%cin\% y}: Case-insensitive %in%

\item \code{x \%ncin\% y}: Case-insensitive %nin%

\item \code{x \%cis\% y}: Case-insensitive ==

\item \code{x \%ncis\% y}: Case-insensitive !=

}}
\examples{
c("apple","pear","coal") \%cin\% c("Banana","Apple","Pear","Cherry")

c("apple","pear","coal") \%ncin\% c("Banana","Apple","Pear","Cherry")

c("APPLE","COAL") \%cis\% c("Apple","Pear")

c("APPLE","COAL") \%ncis\% c("Apple","Pear")

}
\seealso{
[base::`%in%`], [`%nin%`], [base::tolower()]
}
\author{
Sercan Kahveci
}
