% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funs.R
\name{check_bool_scalar}
\alias{check_bool_scalar}
\title{Check whether an object is a valid boolean scalar}
\usage{
check_bool_scalar(x)
}
\arguments{
\item{x}{An R object to check.}
}
\value{
Invisibly returns \code{TRUE} on check pass. Throws a error of class
\code{\link[rlang:rlang_error]{rlang::rlang_error}} on check fail.
}
\description{
This function checks whether an object \code{\link[=is_bool_scalar]{is_bool_scalar()}}. If the check
fails, a formatted error is raised using \code{\link[cli:cli_abort]{cli::cli_abort()}}.
}
\examples{
check_bool_scalar(TRUE)          # passes
check_bool_scalar(FALSE)         # passes
\dontrun{
check_bool_scalar(NA)            # throws error
check_bool_scalar(c(TRUE,FALSE)) # throws error
check_bool_scalar("TRUE")        # throws error
}

}
