% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funs.R
\name{check_char_scalar}
\alias{check_char_scalar}
\title{Check whether an object is a valid character scalar}
\usage{
check_char_scalar(x)
}
\arguments{
\item{x}{An R object to check.}
}
\value{
Invisibly returns \code{TRUE} on check pass. Throws a error of class
\code{\link[rlang:rlang_error]{rlang::rlang_error}} on check fail.
}
\description{
This function checks whether an object \code{\link[=is_char_scalar]{is_char_scalar()}}. If the check
fails, a formatted error is raised using \code{\link[cli:cli_abort]{cli::cli_abort()}}.
}
\examples{
check_char_scalar("hello")      # passes
check_char_scalar("  hi  ")     # passes
\dontrun{
check_char_scalar("")           # error
check_char_scalar("   ")        # error
check_char_scalar(42)           # error
check_char_scalar(c("a", "b"))  # error
check_char_scalar(NA)           # error
check_char_scalar(NULL)         # error
}

}
