% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funs.R
\name{check_logical}
\alias{check_logical}
\title{Check whether an object is a valid boolean scalar}
\usage{
check_logical(x)
}
\arguments{
\item{x}{An R object to check.}
}
\value{
Invisibly returns TRUE. Throws a error of class
\code{\link[rlang:rlang_error]{rlang::rlang_error}}  if \code{x} is not a scalar boolean or an
\code{\link[=is_invalid]{is_invalid()}} object.
}
\description{
This function checks whether an object is a single, non-missing logical
value (\code{TRUE} or \code{FALSE}). If \code{x} \code{\link[=is_invalid]{is_invalid()}} or not a \code{boolean}, it
raises a formatted error using \code{\link[cli:cli_abort]{cli::cli_abort()}}.
}
\examples{
check_logical(TRUE)          # passes
check_logical(FALSE)         # passes
\dontrun{
check_logical(NA)            # throws error
check_logical(c(TRUE,FALSE)) # throws error
check_logical("TRUE")        # throws error
}

}
