% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funs.R
\name{check_numeric_bound}
\alias{check_numeric_bound}
\title{Check whether a variable is within a specified numerical bounds.}
\usage{
check_numeric_bound(
  x,
  lower = 0,
  upper = 1,
  lower_include = TRUE,
  upper_include = TRUE
)
}
\arguments{
\item{x}{The variable to validate.}

\item{lower}{Numerical lower bound. Default 0.}

\item{upper}{Numerical upper bound. Default 1.}

\item{lower_include}{Logical; if \code{TRUE}, the lower bound is inclusive,
otherwise exclusive. Default is TRUE.}

\item{upper_include}{Logical; if \code{TRUE}, the upper bound is inclusive,
otherwise exclusive. Default is TRUE.}
}
\value{
Invisibly returns TRUE if all checks pass; Otherwise, throws
an error of class \link[rlang:rlang_error]{rlang::rlang_error}.
}
\description{
Check whether a variable is within a specified numerical bounds.
}
\examples{
check_numeric_bound(0.5) # within [0, 1]
check_numeric_bound(0, lower_include = TRUE)
check_numeric_bound(1, upper_include = TRUE)

# Exclusive bounds
check_numeric_bound(0.5, lower_include = FALSE, upper_include = FALSE)

# Custom bounds
check_numeric_bound(10, lower = 5, upper = 20)

\dontrun{
# Error cases:
check_numeric_bound(NA)
check_numeric_bound("a")
check_numeric_bound(2, upper = 1)
check_numeric_bound(0, lower_include = FALSE)
}

}
