% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funs.R
\name{check_numeric_scalar}
\alias{check_numeric_scalar}
\title{Check whether an object is a valid numeric scalar}
\usage{
check_numeric_scalar(x)
}
\arguments{
\item{x}{An R object to check.}
}
\value{
Invisibly returns TRUE if \code{x} is a valid numeric scalar. Throws an
error of class \link[rlang:rlang_error]{rlang::rlang_error} otherwise.
}
\description{
This function checks whether an object is a single numeric value (integer or
double), or \code{\link[=is_invalid]{is_invalid()}}. Returns \code{TRUE} invisibly if valid, otherwise raises
a formatted error using \code{\link[cli:cli_abort]{cli::cli_abort()}}.
}
\examples{
check_numeric_scalar(5)       # passes
check_numeric_scalar(0.1)     # passes
\dontrun{
check_numeric_scalar(NA)      # throws error
check_numeric_scalar(NULL)    # throws error
check_numeric_scalar(c(1,2))  # throws error
check_numeric_scalar("5")     # throws error
}

}
