% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{circular.mean}
\alias{circular.mean}
\title{Circular mean}
\usage{
circular.mean(x, mod = 2 * pi, check = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{mod}{The modulus of \code{x}, i.e. where the numbers "loop back" to zero. 
On a clock this is 12, in degrees it is 360, etc.}

\item{check}{Should a check be performed to detect whether there is 
no identifiable mean or median?}

\item{na.rm}{Remove \code{NA}s from \code{x}?}
}
\value{
The circular mean of \code{x}. \code{NA} if \code{check} is \code{TRUE} and
there is no identifiable circular mean.
}
\description{
Circular mean
}
\note{
This should not be confused with the [modular.mean()], which is computed differently.
}
\examples{
# Bedtimes for the past week
bedtimes <- c(23,23,00,01,23,02,03)
circular.mean(bedtimes,mod=24)

# Intuitively, these may require different outcomes, but they have the same circular mean
circular.mean(c(7,11,12),mod=12)
circular.mean(c(8,11,12),mod=12)

# No identifiable circular mean
circular.mean(c(0,pi),mod=2*pi)

}
