% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{clamp}
\alias{clamp}
\alias{clamp0}
\title{clamp}
\usage{
clamp(val, minval = -Inf, maxval = Inf)

clamp0(val, minval = 0, maxval = 1)
}
\arguments{
\item{val}{The vector/matrix to clamp.}

\item{minval}{Minimum value; all lower values are clamped to this value.}

\item{maxval}{Maximum value; all higher values are clamped to this value.}
}
\value{
Clamped vector.
}
\description{
Clamp a numeric vector between a minimum and maximum value.
}
\examples{
clamp(0:10,2,8)
clamp0(rnorm(10))

}
\author{
Sercan Kahveci
}
