% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faststats.R
\name{ClassMetrics}
\alias{ClassMetrics}
\title{Get classification accuracy metrics}
\usage{
ClassMetrics(x, y)
}
\arguments{
\item{x}{The true class memberships}

\item{y}{The predicted class memberships}
}
\value{
A vector with the following accuracy metrics:
\itemize{
\item Accuracy: \code{acc}
\item Chance accuracy: \code{chance_acc}
\item Cohen's Kappa: \code{kappa}
\item Positive predictive value: \code{ppv}
\item Chance positive predictive value: \code{chance_ppv}
\item Kappa of the positive predictive value: \code{kappa_ppv}
\item Negative predictive value: \code{npv}
\item Chance negative predictive value: \code{chance_npv}
\item Kappa of the negative predictive value: \code{kappa_npv}
\item Sensitivity: \code{sens}
\item Specificity: \code{spec}
}
}
\description{
Get classification accuracy metrics
}
\examples{
mymod <- glm(am ~ cyl + disp + hp,
             family="binomial", 
             data = mtcars)

newpred <- as.numeric(predict(mymod, 
                              data=mtcars,
                              type="response") > 0.5)

ClassMetrics(x=mtcars$am, y=newpred)

}
\author{
Sercan Kahveci
}
