% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrangler.R
\name{coalesce.columns}
\alias{coalesce.columns}
\title{Coalesce mutually exclusive column sets into individual columns}
\usage{
coalesce.columns(x, patterns)
}
\arguments{
\item{x}{A \code{data.frame} containing mutually exclusive column sets.}

\item{patterns}{Multiple regex patterns that only match the part of the variable name 
that makes members of the same mutually exclusive variable set unique (i.e. the variant tag).}
}
\value{
A \code{data.frame} with all mutually exclusive column sets coalesced. 
The names of these coalesced columns are based on the original variable names but with 
\code{pattern} removed.
}
\description{
This merges mutually exclusive variables in a \code{data.frame} that 
are distinguished from each other by a suffix or some other part of the variable name.
}
\details{
For example, when a survey was administered in multiple languages, and 
the responses to the same questions in each language are distinguished by a suffix 
like "_EN" or "_FR", this function can be used to 
coalesce these values into individual variables.
}
\examples{
# Create a dataset where different individuals fill in different columns
mydataset <- 
   cbind(id=1:5,
         help_EN=c(1,NA,0,NA,NA),
         help_DE=c(NA,1,NA,1,NA),
         help_FR=c(NA,NA,NA,NA,0),
         success_EN=c(0,NA,0,NA,NA),
         success_DE=c(NA,1,NA,1,NA),
         success_FR=c(NA,NA,NA,NA,1)) |> 
           as.data.frame()
  
coalesce.columns(mydataset,c("_EN","_DE","_FR"))

}
