% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{cols2ids}
\alias{cols2ids}
\title{Transform all data.frame columns into ordered IDs}
\usage{
cols2ids(x)
}
\arguments{
\item{x}{A \code{data.frame}.}
}
\value{
\code{x} with each column replaced by integer IDs.
}
\description{
Transforms every column into a factor and then into an integer.
}
\examples{
cols2ids(mtcars)

}
\author{
Sercan Kahveci
}
