% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{colVars}
\alias{colVars}
\alias{rowVars}
\title{Compute column and row variances}
\usage{
colVars(x, na.rm = T)

rowVars(x, na.rm = T)
}
\arguments{
\item{x}{an input matrix of data.frame}

\item{na.rm}{Logical indicating whether NA values should be omitted before variance computation}
}
\description{
Compute column and row variances
}
\examples{
colVars(WorldPhones)
rowVars(WorldPhones)
}
