% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{combobulate}
\alias{combobulate}
\title{Get all possible combinations of strings}
\usage{
combobulate(...)
}
\arguments{
\item{...}{Character vectors to combobulate.}
}
\value{
A character vector.
}
\description{
\code{combobulate()} returns all possible combinations of the provided character strings, each combination merged into a single string.
}
\examples{
combobulate("Hello ",c("Sir","Madam"),", ",c("may I take your order?","what shall it be?"))
# [1] "Hello Sir, may I take your order?"    
# [2] "Hello Madam, may I take your order?"   
# [3] "Hello Sir, what shall it be?"  
# [4] "Hello Madam, what shall it be?"
}
