% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{comboTable}
\alias{comboTable}
\title{Generate a matrix of combinations of values}
\usage{
comboTable(...)
}
\arguments{
\item{...}{Character vectors, named or unnamed, or unquoted names of named arguments. 
Character vectors will be used to generate a matrix where each row represents a unique combination
of all values, akin to \code{expand.grid()}. Arguments which are unquoted names of named arguments
will become copies of the column generated by the eponymous named character vector.}
}
\value{
A matrix.
}
\description{
Generate a matrix of combinations of values
}
\examples{
hh<-c("a","b")
comboTable(a=letters[1:3], b=2,a,b,c=c("e","f"),d,c,d=hh,"huh",a,hh)
}
