% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{cons.rank}
\alias{cons.rank}
\title{Produce consecutive ranks}
\usage{
cons.rank(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
An integer vector of the same length as \code{x}.
}
\description{
Ranks a vector such that ranks are always 1 apart if they succeed each other
or 0 apart if they are tied.
}
\examples{

testvec <- c(0,2,3,3,7,9,9,0)
cons.rank(testvec)

}
\author{
Sercan Kahveci
}
