% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skMisc.R
\name{CorrCrunch}
\alias{CorrCrunch}
\title{Analyse the robustness of a correlation}
\usage{
CorrCrunch(x, y, verbose = F)
}
\arguments{
\item{x, y}{Numeric vectors to correlate.}

\item{verbose}{if TRUE, prints verbose output.}
}
\value{
A list containing the number of cases that need to be removed to flip the sign of the correlation coefficient; 
the proportion removed cases in the data; and a data.frame without these cases.
}
\description{
\code{CorrCrunch()} computes the minimum number of cases that need to be removed from a dataset to flip the sign of a correlation coefficient.
This can be useful in distinguishing genuine correlations from spurious findings that hinge on one or two outliers.
Cases are removed iteratively; in each iteration the case that maximally shrinks the correlation coefficient is removed.
}
\examples{
CorrCrunch(mtcars$mpg,mtcars$wt)
#Holdout needed to flip the sign: 19 (63.33\%)
#Final r: 0.01181141
}
