% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faststats.R
\name{CorTable}
\alias{CorTable}
\title{Generate a correlation table}
\usage{
CorTable(
  x,
  method = c("pearson", "spearman"),
  holdout.goal = c("nsig", "flip"),
  alpha = 0.05
)
}
\arguments{
\item{x}{A matrix or data.frame of which the columns will be correlated with each other}

\item{method}{Whether to use pearson or spearman correlations}

\item{holdout.goal}{When computing the holdout statistic with [cor.holdout()], what should
the goal be - non-significance ("nsig") or a flip of the sign ("flip")?}

\item{alpha}{Alpha level for non-significance testing in case \code{holdout.goal="nsig"}}
}
\value{
A list with 5 items - 4 matrices 
(r - correlation, n = sample size, p = p-value, h = holdout value) 
and a list of parameters, such as the correlation type.
NA values among the holdout values indicate that computation failed, e.g. 
due to insufficient variance or an inability to achieve the holdout goal
without depleting the sample size.
}
\description{
This computes correlations from the imputed data, yielding a correlation matrix
as well as matrices for sample size, p-values, and holdout values (computed with
[cor.holdout()]). The latter is a robustness check for the obtained correlation,
denoting how many observations need to be removed before a certain goal is achieved,
such as non-significance, or a flip of the sign of the correlation coefficient
(represented by the argument \code{holdout.goal}).
}
\details{
The output of this object can be printed with [print.CorTable()].
}
\examples{
test<-CorTable(mtcars,holdout.goal="flip")
print(test)

test<-CorTable(mtcars,holdout.goal="nsig")
print(test,type="r/h")

}
\author{
Sercan Kahveci
}
