% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_article_template.R
\name{create_article_template}
\alias{create_article_template}
\title{Create a Structured Article Template for a Specified Journal}
\usage{
create_article_template(journal_name)
}
\arguments{
\item{journal_name}{A character string specifying the journal for which the article
template will be created. It must be one of the supported journals listed by
\code{rticles::journals()}.}
}
\value{
This function does not return a value. It creates a directory structure for
the article template, including the necessary \code{Rmd} files for sections and the abstract.
}
\description{
This function generates an article template for the specified journal using the
\code{rmarkdown::draft} function from the \code{rticles} package. It checks whether
the journal is supported, creates the necessary directories and files, and formats
the \code{<journal_name>_article.Rmd} file by adding section chunks and an abstract.
}
\examples{
\dontrun{
create_article_template("arxiv")
}
}
