% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{df.init}
\alias{df.init}
\title{Initiate an empty data frame}
\usage{
df.init(x, nrow = 0)
}
\arguments{
\item{x}{A character vector of column names.}

\item{nrow}{Number of rows (defaults to 0).}
}
\value{
A data.frame filled with \code{NA}.
}
\description{
Initiate an empty data frame
}
\examples{
test <- df.init(c("A","B","C"))

}
\author{
Sercan Kahveci
}
