% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceNetworks.R
\name{EBIC.glmnet}
\alias{EBIC.glmnet}
\title{Extended Bayesian Information Criterion for \code{glmnet} objects}
\usage{
EBIC.glmnet(x, gamma)
}
\arguments{
\item{x}{The \code{glmnet} model.}

\item{gamma}{The gamma parameter. This should be a value between 0 and 1.
At a gamma of 0, EBIC is equivalent to BIC, while at higher values, it is
more strict on the number of predictors modelled.}
}
\value{
The EBIC of the model.
}
\description{
Extended Bayesian Information Criterion for \code{glmnet} objects
}
