% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4tools.R
\name{ExpandFormula}
\alias{ExpandFormula}
\title{Parse a lme4 formula and return all main effects and interactions as separate terms}
\usage{
ExpandFormula(form)
}
\arguments{
\item{form}{A formula to be expanded.}
}
\value{
The same formula, but with all interactions and mai neffects as separate terms
}
\description{
Parse a lme4 formula and return all main effects and interactions as separate terms
}
\examples{
ExpandFormula(rt ~ pull * target + (pull * target | subjectid))
#> rt ~ pull + target + pull:target + (pull + target + pull:target | subjectid)

ExpandFormula(rt ~ pull * target + (pull * target || subjectid))

}
