% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4tools.R
\name{ExtractRandomTerms}
\alias{ExtractRandomTerms}
\title{Extract random terms from a lme4 formula}
\usage{
ExtractRandomTerms(form)
}
\arguments{
\item{form}{A formula}
}
\value{
A named list containing character vectors with random terms; names are group variables.
}
\description{
Extract random terms from a lme4 formula
}
\examples{
ExtractRandomTerms(grade ~ ChildIQ * TeacherSkill * SchoolType + (ChildIQ * TeacherSkill | School))
#$School
#[1] "ChildIQ"              "TeacherSkill"         "ChildIQ:TeacherSkill"
}
