% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faststats.R
\name{faststats}
\alias{faststats}
\alias{zerodiff}
\alias{twodiff}
\alias{twodiff2}
\alias{npr.zerodiff}
\alias{npr.twodiff}
\alias{twocor}
\title{Statistics functions with concise output}
\usage{
zerodiff(x)

twodiff(form, data, paired = FALSE)

twodiff2(x, y, paired = FALSE)

npr.zerodiff(x)

npr.twodiff(form, data)

twocor(x, y, method = "pearson")
}
\arguments{
\item{x, y}{Vectors to compare or correlate}

\item{form}{A formula where the left size is the values to be compared and
the right value is a binary variable defining the groups to compare}

\item{data}{a \code{data.frame}}

\item{paired}{whether to perform a paired or unpaired test}

\item{method}{The correlation type; can be pearson, spearman, or kendall}
}
\description{
These were mainly created because the base stats functions
produce overly verbose output without effect sizes included.
}
\examples{
baseline <- rnorm(50)
deviation <- rnorm(50,m=1)

zerodiff(deviation)

}
