% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetworks.R
\name{findloops}
\alias{findloops}
\title{Detect feedback loops in a network}
\usage{
findloops(mat, maxnodes, minnodes = 2)
}
\arguments{
\item{mat}{A matrix of edge weights}

\item{maxnodes, minnodes}{Maximum and minimum number of nodes
permitted to be involved in the loop for it to be included in the output.}
}
\value{
A list with:
\itemize{
\item A list of character vectors, each being a loopderived from the network
\item A \code{data.frame} with one loop per row, describing
\itemize{
\item a character description of the involved nodes and
the sign of the edges between them
\item The product of all edge weights,
indicating whether it is a positive or negative feedback loop
\item The number of nodes involved
}
\item A list of matrices, each of the same size as \code{mat} and
retaining only the edges involved in the loop
}
}
\description{
Detects feedback loops in a network.
}
\details{
Note: this has not yet been tested and may not behave as intended.
}
\author{
Sercan Kahveci
}
