% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{findQuantile}
\alias{findQuantile}
\title{Find quantiles of vector}
\usage{
findQuantile(x, n)
}
\arguments{
\item{x}{A numeric vector to replace with the quantile of its values.}

\item{n}{The number of quantiles to divide it into.}
}
\value{
A vector with values replaced by their quantile membership.
}
\description{
Will replace vector values with their quantile.
}
\examples{
a<-rnorm(100)
findQuantile(a,5)

}
\author{
Sercan Kahveci
}
