% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4tools.R
\name{FindTopTerms}
\alias{FindTopTerms}
\title{Find all model terms that are not moderated by a higher-order interaction}
\usage{
FindTopTerms(form)
}
\arguments{
\item{form}{a formula}
}
\value{
A character vector containing all model terms that are not moderated by a higher-order interaction.
}
\description{
Find all model terms that are not moderated by a higher-order interaction
}
\examples{
FindTopTerms(speed ~ skill + weight * friction)
#[1] "skill"           "weight:friction"
}
