% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_stat}
\alias{format_stat}
\title{Format statistics appropriately}
\usage{
format_stat(
  x,
  digits = 2,
  type = c("default", "p", "quotient"),
  sign.positive = FALSE
)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{How many digits to round the number to.}

\item{type}{Type of statistic formatting rule to apply. 
"p" prints values rounded to zero as <.001, and 
"quotient" prints values <1 as 1/x where x is a value above 1 
(e.g. in case of Bayes factors).}

\item{sign.positive}{If \code{TRUE}, adds a + to every positive number.}
}
\value{
A character vector of formatted numbers.
}
\description{
This formats numbers according to APA.
}
\details{
All leading zeros are also dropped.
}
\examples{
format_stat(0.12345678)

# Proper printing of p-values
format_stat(0.0004,type="p")

# Printing of quotients where the range of values between 1 and 0
# should be considered equal to that between 1 and infinity
format_stat(0.05,type="quotient")

}
\author{
Sercan Kahveci
}
