% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetworks.R
\name{getConnectedNodes}
\alias{getConnectedNodes}
\alias{pruneUnconnectedNodes}
\title{Get all connected nodes in a network}
\usage{
getConnectedNodes(
  mat = NULL,
  net = NULL,
  node = NULL,
  ignore.selfloops = TRUE,
  ignore.negative = FALSE,
  components = c("all", "largest")
)

pruneUnconnectedNodes(
  mat = NULL,
  node = NULL,
  ignore.selfloops = TRUE,
  ignore.negative = FALSE,
  components = c("all", "largest")
)
}
\arguments{
\item{mat}{An adjacency matrix.}

\item{net}{A \code{qgraph} object. 
One of \code{mat} or \code{net} must be given.}

\item{node}{A node name. If given, the function only returns nodes that are
connected to this node, directly or indirectly.}

\item{ignore.selfloops}{Should self-loops be ignored when determining
if a node is unconnected?}

\item{ignore.negative}{Should negative edges be ignored when determining
if a node is unconnected?}

\item{components}{If \code{"largest"}, this function only returns nodes 
that are part of the largest component of the network. If \code{"all"},
no such subsetting is performed.}
}
\value{
A vector of node names that are connected to other nodes according to
the rules provided in this function's arguments.
}
\description{
Get all connected nodes in a network
}
\author{
Sercan Kahveci
}
