% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetworks.R
\name{getNodePaths}
\alias{getNodePaths}
\title{Get all edges leading to/from a node}
\usage{
getNodePaths(mat, from = NULL, to = NULL)
}
\arguments{
\item{mat}{An adjacency matrix}

\item{from, to}{Node names between which the shortest path must be computed.
At least one of these must be given. 
If only \code{from} is given, then this returns 
the shortest paths from this node to all other nodes, 
and if only \code{to} is given, then this returns 
the shortest paths to this node from all other nodes.}
}
\value{
The same adjacency matrix, but the only non-null values are those edges
on the shortest path from the \code{from} node and/or to the \code{to} node.
}
\description{
Get all edges leading to/from a node
}
\author{
Sercan Kahveci
}
