% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{ggplot.themes}
\alias{ggplot.themes}
\alias{theme_pecher}
\alias{theme_apa}
\title{extra \code{ggplot2} themes}
\usage{
theme_pecher()

theme_apa()
}
\description{
A number of ready-to-use ggplot2 themes for scientific manuscripts
}
\section{Functions}{
\itemize{
\item \code{theme_pecher()}: Based on the plot design style of prof. Diane Pecher.

\item \code{theme_apa()}: Theme appropriate for APA manuscripts

}}
\examples{
ggplot(mtcars,aes(x=mpg,y=wt)) + geom_point() + theme_pecher()

ggplot(mtcars,aes(x=mpg,y=wt)) + geom_point() + theme_apa()

}
