% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{gini}
\alias{gini}
\title{Gini coefficient}
\usage{
gini(x, unbiased = FALSE)
}
\arguments{
\item{x}{A vector to compute the Gini coefficient of}

\item{unbiased}{Whether to multiply by n/(n-1) to ensure generalizability to the population-level.}
}
\value{
The Gini coefficient of the vector. 
\code{NA} if the vector contains a negative, zero, or NA value.
}
\description{
Computes the Gini coefficient, an index of inequality.
}
\examples{
gini(state.x77[,"Income"])

# Completely equal gives 0
gini(c(100,100,100))

# Outliers increase the value
gini(c(100,100,1000))

}
\seealso{
\href{https://www.statsdirect.com/help/nonparametric_methods/gini.htm}{Source utilized}

Mills, J.A., Zandvakili, A. (1997). Statistical inference via bootstrapping for measures of inequality. 
Journal of Applied Econometrics, 12, 133-150. https://www.jstor.org/stable/2284908
}
