% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Negative %in%
Returns which values in the left vector are not in the right vector.}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{Values whose presence will be checked for in \code{table}}

\item{table}{Values that will yield a \code{FALSE} if they exist in \code{x}}
}
\value{
A logical vector indicating whether each value of \code{x} lacks a match in \code{table}
}
\description{
Negative %in%
Returns which values in the left vector are not in the right vector.
}
\examples{
(1:5) \%nin\% (1:3)

}
