% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{hedgesg}
\alias{hedgesg}
\title{Hedges' g}
\usage{
hedgesg(x, y = NULL, paired = FALSE)
}
\arguments{
\item{x, y}{Numeric vectors. \code{y} can be omitted if 
the effect size should be computed for the difference from zero.}

\item{paired}{Are the values in \code{x} and \code{y} paired? 
Then Hedges' g will be computed for \code{x-y}.}
}
\value{
Hedges' g, positive if x > y.
}
\description{
A measure of effect size similar to Cohen's d with a more accurate 
pooled standard deviation of the contrasted groups.
}
\examples{
# Difference from zero
hedgesg(x=rnorm(100,m=3))

# Difference from another group
hedgesg(x=rnorm(100,m=0),y=rnorm(100,m=3))

# Difference from paired value
a <- rnorm(100,m=3)
b <- a+rnorm(100,m=3)
hedgesg(x=a,y=b)

}
