% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hilight}
\alias{hilight}
\title{Plot highlighted text}
\usage{
hilight(x, y, s, bg = "yellow")
}
\arguments{
\item{x}{x position}

\item{y}{y position}

\item{s}{text}

\item{bg}{highlight color}
}
\description{
Plot highlighted text
}
\examples{
plot(mtcars$mpg,mtcars$wt,col=mtcars$cyl)
hilight(27,2.5,"Light and\nefficient")
hilight(17,4.5,"Heavy and\ninefficient")

}
\author{
Sercan Kahveci
}
