% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkapply.R
\name{ichunk}
\alias{ichunk}
\title{Iterate over chunks of a vector}
\usage{
ichunk(chunksize, count = NULL)
}
\arguments{
\item{chunksize}{How long should each sequence be?}

\item{count}{What should be the maximum value returned by the iterator? 
Above this value, the iterator will return \code{NULL}.}
}
\value{
An iterator function.
}
\description{
This returns an iterator that counts from one, returning integer sequences 
of a pre-specified length with each iteration up until an optional maximum value.
}
\note{
This does not return an iterator in the manner used by the iterators package,
since those objects currently misbehave in an Rstudio live session.
}
\examples{
myit <- ichunk(chunksize=10,count=15)
myit()
myit()
myit()

}
