% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{interweave}
\alias{interweave}
\title{Interweave vectors}
\usage{
interweave(...)
}
\arguments{
\item{...}{Any number of vectors of equal length or length 1.}
}
\value{
The input vectors interwoven into one.
}
\description{
Interweave vectors, such that the result features the first value from vector A, 
then the first value from vector B, then the second value from vector A, etc.
}
\examples{
# Interweave 3 equal-length vectors 
interweave(1:3, 4:6, 7:9)

# Interweave vector with single value
interweave(1:5, 0)

}
\author{
Sercan Kahveci
}
