% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_funs.R
\name{is_invalid}
\alias{is_invalid}
\title{Test whether an object is considered invalid}
\usage{
is_invalid(x)
}
\arguments{
\item{x}{An R object to check.}
}
\value{
A logical scalar: TRUE if the object is invalid, FALSE otherwise.
}
\description{
This function tests whether a R object is considered invalid. An object is
invalid if it is either of the followings,
\itemize{
\item \link[base:missing]{base::missing},
\item \code{NULL},
\item has zero length,
\item an object with class \code{try-error},
\item an object with class \code{simpleError},
\item a list whose elements are all invalid (see details) or
\item an atomic vector whose elements are all \code{NA}.
}
}
\details{
Lists are checked recursively: a list is invalid only when all of its
elements are invalid.
}
\examples{
is_invalid(NULL)                 # TRUE
is_invalid(character(0))         # TRUE
is_invalid(c(NA, NA))            # TRUE
is_invalid(c(1, NA))             # FALSE

is_invalid(list(NA, NULL))       # TRUE
is_invalid(list(NA, 5))          # FALSE

# try-error example
err1 <- try(
  stop("err"), silent = TRUE
)
is_invalid(err1)                 # TRUE

# simpleError example
err2 <- simpleError("oops")
is_invalid(err2)                 # TRUE

is_invalid(5)                    # FALSE
is_invalid("txt")                # FALSE

}
