% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{LeaveOneOut}
\alias{LeaveOneOut}
\alias{loo_sum}
\alias{loo_mean}
\alias{loo_var}
\alias{loo_sd}
\alias{loo_cov}
\alias{loo_cor}
\title{Fast leave-one-out statistics}
\usage{
loo_sum(x)

loo_mean(x)

loo_var(x)

loo_sd(x)

loo_cov(x, y)

loo_cor(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\value{
A numeric vector of equal length to \code{x} but with each value representing the
requested statistic that would be obtained if the element
in the same position in \code{x} and \code{y} were left out.
}
\description{
These functions return a vector where each element represents
a statistic applied to that vector but with that particular element left out
(i.e., leave-one-out statistics).
}
\details{
The functions are optimized and should run faster than
if you would actually leave a value out and recalculate the statistic.
}
\section{Functions}{
\itemize{
\item \code{loo_sum()}: leave-one-out sum

\item \code{loo_mean()}: leave-one-out mean

\item \code{loo_var()}: leave-one-out variance

\item \code{loo_sd()}: leave-one-out standard deviation

\item \code{loo_cov()}: leave-one-out covariance

\item \code{loo_cor()}: leave-one-out Pearson correlation

}}
\examples{
testvec <- runif(2000)
testdata <- MASS::mvrnorm(30,mu=c(0,1),Sigma=100*matrix(c(1,.5,.5,1),nrow=2))

# Sum
loosum <- loo_sum(testvec)
loosum2 <- sapply(seq_along(testvec),\(i)sum(testvec[-i]))
max(abs(loosum-loosum2)) # Largest error

# Mean 
loomean <- loo_mean(testvec)
loomean2 <- sapply(seq_along(testvec),\(i)mean(testvec[-i]))
max(abs(loomean-loomean2))

# Variance
loovar <- loo_var(testvec)
loovar2 <- sapply(seq_along(testvec),\(i)var(testvec[-i]))
max(abs(loovar-loovar2))

# SD
loosd <- loo_sd(testvec)
loosd2 <- sapply(seq_along(testvec),\(i)sd(testvec[-i]))
max(abs(loosd-loosd2))

# Covariance
loocov <- loo_cov(testdata[,1],testdata[,2])
loocov2 <- sapply(seq_len(nrow(testdata)),\(i)cov(testdata[-i,1],testdata[-i,2]))
max(abs(loocov-loocov2))

# Pearson correlation
loocor <- loo_cor(testdata[,1],testdata[,2])
loocor2 <- sapply(seq_len(nrow(testdata)),\(i)cor(testdata[-i,1],testdata[-i,2]))
max(abs(loocor-loocor2))

}
\author{
Sercan Kahveci
}
