% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{LevenshteinDistance}
\alias{LevenshteinDistance}
\title{Levenshtein distance}
\usage{
LevenshteinDistance(source, target)
}
\arguments{
\item{source, target}{Strings to be compared.}
}
\value{
The Levenshtein distance between the two strings.
}
\description{
Counts the number of single character deletions, insertions, and substitutions 
that need to be performed to turn the source string into the target string.
}
\examples{
LevenshteinDistance("Yoghurt","Youtube")
}
