% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic.R
\name{logical-na-handlers}
\alias{logical-na-handlers}
\alias{NA2TRUE}
\alias{NA2FALSE}
\title{Coerce NA values to \code{TRUE} or \code{FALSE}}
\usage{
NA2TRUE(x)

NA2FALSE(x)
}
\arguments{
\item{x}{A logical vector.}
}
\value{
\code{x} with \code{NA} values replaced with 
\code{TRUE} or \code{FALSE}.
}
\description{
Coerce NA values to \code{TRUE} or \code{FALSE}
}
\examples{
fruits <- c("apples","pears",NA,"cherries")
NA2TRUE(fruits != "apples")
NA2FALSE(fruits == "pears")

}
\author{
Sercan Kahveci
}
