% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{logit.weightfun}
\alias{logit.weightfun}
\title{Downweight outliers}
\usage{
logit.weightfun(x, mean = mean(x), s = sd(x), sdist = 3,
  taper = 10, scale = c("max", "norm"), min = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{mean}{An optional mean of the vector}

\item{s}{An optional standard deviation of the vector}

\item{sdist}{The number of standard deviations beyond which values should be downweighted}

\item{taper}{A number indicating how strongly values exceeding the standard deviation should taper off}

\item{scale}{How the weight vector should be scaled: "norm" sets the sum to 1, "max" sets the maximum to 1.}

\item{min}{A minimum weight.}
}
\value{
A numeric vector of weights
}
\description{
Computes weights; trials within certain bounds of the mean receive the maximum weight while trials
outside these bounds are downweighted to 0 or an optional minimum.
}
