% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrangler.R
\name{match.merge}
\alias{match.merge}
\title{Match and merge two data.frames by ID and consecutive date}
\usage{
match.merge(
  x,
  y,
  idvar,
  datevar,
  by = NULL,
  keep.unmatched = c("all", "x", "y", "none"),
  suffixes = c(".x", ".y")
)
}
\arguments{
\item{x, y}{The \code{data.frame}s to merge.}

\item{idvar}{Name of the variable holding participant/session IDs in both x and y.}

\item{datevar}{Name of the variable holding dates in both x and y.}

\item{by}{Any other variables to match by.}

\item{keep.unmatched}{When no match is found for a row, should it be kept or removed?}

\item{suffixes}{Character strings to append to the end of variable names that occur in both
\code{x} and \code{y}, to make them unique in the resulting \code{data.frame}.}
}
\value{
A merged \code{data.frame} where rows of \code{x}
are merged with rows of \code{y} that match by session id and where
the date of each row from \code{x} directly precedes the 
date of the matched row from \code{y}.
}
\description{
the rows from \code{y} are matched with rows from \code{x} that 
share the same ID and directly precede them in date.
}
\details{
This is useful for matching participant data from different datasets,
where one dataset is always collected after another. It is assumed here that
\code{x} is always administered before \code{y}.

This matches using [match.pps()] and merges using [dplyr::full_join()].
}
\examples{
x <- data.frame(  id=c(1,2,2,3,3,4,5),
                date=c(1,1,3,1,3,1,9),
               info=letters[1:7])
y <- data.frame(  id=c(1,1,2,3,3,4),
                date=c(2,0,2,4,2,0),
                data=LETTERS[1:6])
match.merge(x=x,y=y,idvar="id",datevar="date",keep.unmatched="all")
match.merge(x=x,y=y,idvar="id",datevar="date",keep.unmatched="none")

}
\author{
Sercan Kahveci
}
