% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{modular.mean}
\alias{modular.mean}
\alias{modular.median}
\title{Mean or median of a modular quantity}
\usage{
modular.mean(x, mod, check = TRUE, na.rm = FALSE)

modular.median(x, mod, check = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector of modular quantities to be aggregated}

\item{mod}{The modulus of \code{x}, i.e. where the numbers "loop back" to zero. 
On a clock this is 12, in degrees it is 360, etc.}

\item{check}{Should a check be performed to detect whether there is 
no identifiable mean or median?
Set to \code{FALSE} when this is not expected to occur, since this check can be costly.}

\item{na.rm}{Should \code{NA} values be removed? Defaults to \code{TRUE}.}
}
\value{
\code{modular.mean()} and \code{modular.median()} respectively find 
the value with the smallest squared or absolute distance 
to all values in \code{x}, like the regular mean and median do.
}
\description{
Mean or median of a modular quantity
}
\details{
These functions should not be confused with the [circular.mean()], which is computed
in a different way.
}
\examples{
modular.mean(c(0,1,2,10),mod=12)

# Situations where no modular mean exists (points are evenly divided around the circle)
modular.mean(c(0,3,6,9),mod=12)
modular.mean(c(0,1,3,4,6,7,9,10),mod=12)

}
