% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{multimerge}
\alias{multimerge}
\title{Merge Multiple Data Frames}
\usage{
multimerge(x, ...)
}
\arguments{
\item{x}{a list of data frames}

\item{...}{all other arguments for \code{merge} can be provided here}
}
\value{
A single, merged \code{data.frame}
}
\description{
This function makes calls to \code{merge()} to merge every other dataset 
with the one next to it, repeating until only one dataset remains.
}
\examples{
#generate test data
testlist<-list()
lsize<-50
for(i in 1:lsize){
  testlist[[i]]<-data.frame(key=sample(1:500,100),
                            junk=letters[sample(1:26,100,replace=TRUE)])
  colnames(testlist[[i]])[2]<-paste0("info",i)
}
multimerge(testlist,by="key",all=TRUE)

}
