% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{NONE2NA}
\alias{NONE2NA}
\title{Convert empty vectors to NA}
\usage{
NONE2NA(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
\code{NA} if \code{x} is length 0, else \code{x} itself.
}
\description{
This converts vectors of length 0 to \code{NA}.
}
\examples{
NONE2NA(NULL)
NONE2NA(logical(0))
NONE2NA(1:5)

}
