% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{nthWeekdayOfMonth}
\alias{nthWeekdayOfMonth}
\title{Convert dates to Nth weekday of the month values
Computes which weekday of the month each date represents. 
E.g., for each "second monday of the month", it gives 2.}
\usage{
nthWeekdayOfMonth(x)
}
\arguments{
\item{x}{Date(s) to convert}
}
\value{
A vector of Nth weekday of the month values
}
\description{
Convert dates to Nth weekday of the month values
Computes which weekday of the month each date represents. 
E.g., for each "second monday of the month", it gives 2.
}
\examples{
nthWeekdayOfMonth("2000-04-08")

}
