% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{OLcrunch}
\alias{OLcrunch}
\title{Crunch Outliers}
\usage{
OLcrunch(x, DS = 3, hardlimit = NULL)
}
\arguments{
\item{x}{Numeric vector to remove outliers from}

\item{DS}{A positive numeric value. If value exceeds this many standard deviations, it is counted as an outlier}

\item{hardlimit}{A numeric vector with two values. If set, values below the first value and above the second will be counted as outliers, and means/standard deviations will be computed from values within these bounds only.}
}
\value{
Vector with outlying values set to NA
}
\description{
Crunch Outliers
}
