% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{pair}
\alias{pair}
\title{Create unique pairs}
\usage{
pair(...)
}
\arguments{
\item{...}{two or more vectors of equal length}
}
\value{
a character vector consisting of all input vectors concatenated term-by-term and in alphabetic order.
}
\description{
Combines vectors such that unique unordered sets are derived from the vectors' cross sections.
}
\examples{
pair(1:4,4:1)
#[1] "1-4" "2-3" "2-3" "1-4"
}
