% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{plotmat}
\alias{plotmat}
\title{Plot matrix as heatmap}
\usage{
plotmat(x, text = FALSE, plot = TRUE, ...)
}
\arguments{
\item{x}{A matrix.}

\item{text}{Whether to print the values of the matrix as text (defaults to FALSE).}

\item{plot}{Whether to plot the resulting ggplot object (defaults to TRUE).}

\item{...}{Ignored.}
}
\value{
Invisibly returns the ggplot object for further modification.
}
\description{
Plot matrix as heatmap
}
\examples{
# Create a matrix to plot
mymat <- matrix(rnorm(100), nrow=10)
mymat <- mymat + t(mymat)
mymat[cbind(sample(c(1:10)), sample(c(1:10)))] <- NA
colnames(mymat) <- rownames(mymat) <- sample(letters[1:10])

plotmat(mymat)

}
\author{
Sercan Kahveci
}
