% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faststats.R
\name{print.CorTable}
\alias{print.CorTable}
\title{Print a \code{CorTable} object}
\usage{
\method{print}{CorTable}(x, type = c("full", "r/p", "r/h"), alpha = 0, digits = 2, ...)
}
\arguments{
\item{x}{a \code{CorTable} object}

\item{type}{The type of table to print.
\itemize{
\item \code{"full"} Separately print the correlation, sample size, p-value, and holdout matrices
\item \code{"r/p"} Print a single matrix with p-values in the upper triangle and
correlations in the lower triangle
\item \code{"r/h"} Print a single matrix with holdout values in the upper triangle and
correlations in the lower triangle
}}

\item{alpha}{Correlations with a p-value below this value will be highlighted
with an asterisk.}

\item{digits}{Digits to round the values by. p-values will be rounded by this value +1}

\item{...}{Ignored}
}
\value{
The printed character matrix is silently returned.
}
\description{
Print a \code{CorTable} object
}
\examples{
test<-CorTable(mtcars,holdout.goal="flip")
print(test)

test<-CorTable(mtcars,holdout.goal="nsig")
print(test,type="r/h")

}
