% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{quantize}
\alias{quantize}
\title{Quantize vector}
\usage{
quantize(x, n)
}
\arguments{
\item{x}{A numeric vector to quantize}

\item{n}{The number of quantiles to divide it into}
}
\value{
A vector with values replaced by their quantile membership
}
\description{
Will replace vector values with their quantile.
}
\examples{
a<-rnorm(100)
quantize(a,5)

}
