% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratelimits.R
\name{RateLimit}
\alias{RateLimit}
\alias{defineRateLimit}
\alias{awaitRateLimit}
\title{Function call rate limiting}
\usage{
defineRateLimit(period, times, padding = 0)

awaitRateLimit(queue)
}
\arguments{
\item{period}{Time period in seconds within which the rate of function calls must be limited.}

\item{times}{Maximum number of times this queue can be utilized within the designated \code{period}.}

\item{padding}{Additional amount of time in seconds to wait until the script continues, 
when the rate limit has been exceeded.}

\item{queue}{The rate limit object.}
}
\description{
Function call rate limiting
}
\examples{
h<-defineRateLimit(period=60,times=30)
for(i in 0:30){
  message(i)
  awaitRateLimit(h)
}

}
