% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{read.csv.folder}
\alias{read.csv.folder}
\title{Read and merge all .csv files in a folder}
\usage{
read.csv.folder(folder = "./", readfunc = list(read.csv, read.csv2,
  read.table))
}
\arguments{
\item{folder}{path to a folder}

\item{readfunc}{list of functions that will be used to read the files; if the first function fails, the second function will be used, etc.}
}
\value{
A data.frame containing all merged .csv files
}
\description{
Read and merge all .csv files in a folder
}
