% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faststats.R
\name{removeOLs}
\alias{removeOLs}
\title{Remove outlying observations from a data.frame or matrix}
\usage{
removeOLs(.tbl, olvars = NULL, groups = NULL, s = 3, make.na = FALSE)
}
\arguments{
\item{.tbl}{A \code{data.frame} or matrix to exclude outliers from}

\item{olvars}{Names or numeric index of the variables to detect outliers in. 
If \code{NULL}, all variables will be checked for outliers.}

\item{groups}{(optional) name or numeric index of the variable identifying groups of observations; 
outlier detection will be performed separately per group.}

\item{s}{If a value deviates more SDs from the mean than this value, it is marked as an outlier}

\item{make.na}{If \code{FALSE} (default), excludes all rows that have an outlier in
at least one variable in \code{olvars} (listwise).
If \code{TRUE}, the function instead turns the individual outlying values into \code{NA},
and does not exclude any rows.}
}
\value{
The input \code{data.frame} or matrix with outliers excluded.
}
\description{
Outliers are defined as values deviating more than Xstandard deviations (SDs) from the mean.
}
\details{
This does not detect any outliers in groups with less than 3 non-NA observations.
}
\examples{
# Standard deviation limits can be set with argument s
removeOLs(mtcars, olvars=c("mpg", "disp", "hp"))
removeOLs(mtcars, olvars=c("mpg", "disp", "hp"), s=1)

# Replace OLs with NA with argument make.na
testdata <- mtcars
testdata$mpg[1] <- 40
testdata$hp[2] <- 500
removeOLs(testdata, olvars=c("mpg", "disp", "hp"), groups="vs", make.na=T)

# Also works on matrices
testmat <- matrix(rnorm(1000), ncol=5)
testmat[cbind(sample(1:200,5),1:5)]<-1000
removeOLs(testmat)

}
\seealso{
[vec.removeOLs()] for the same outlier exclusion applied to a single vector.
}
