% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{retry}
\alias{retry}
\title{Retry running a function until it succeeds}
\usage{
retry(expr, n = 5)
}
\arguments{
\item{expr}{An expression to try to execute.}

\item{n}{Maximum number of times to try to run the expression.}
}
\value{
The output of the expression.
}
\description{
An expression is executed using [base::try()] and re-run until
it raises no more errors or until a maximum number of evaluations is reached.
In the latter case, it raises an error of its own.
}
\details{
This was primarily developed to robustly query websites. Sometimes
a query fails or produces output which raises an error; 
this function issues a retry in such a case.
}
\examples{
k <- 0
retry({k <<- k+1; if(k<4){ stop()}})


}
