% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{rotate2front}
\alias{rotate2front}
\title{Rotate a specific value to the front of vectors}
\usage{
rotate2front(x, first = NULL, matches.only = FALSE)
}
\arguments{
\item{x}{List of vectors or a single vector to align through rotation.}

\item{first}{Optional vector indicating which values would be preferred as first value,
in order of preference.
If left blank, this is set to the unique values in \code{x} in order of occurrence.}

\item{matches.only}{Should vectors only be rotated if at least one value
has a match in \code{first}? Default is \code{FALSE}. If set to \code{FALSE},
then vectors without a match will be rotated such that the first value is the first
that occurs in \code{x}.}
}
\value{
If \code{x} was a list, this returns
a list with all constituent vectors rotated in accordance with the algorithm.
If \code{x} was a vector, this returns
a vector rotated in accordance with the algorithm.
}
\description{
This function rotates vectors in a list such that the first value in each vector
is the first value in \code{first} that occurs in that vector.
}
\details{
This can be used to identify cyclic vectors that are duplicated
with \code{duplicated(alignchains(x))}.
}
\examples{
vecs <- list(a=c(5,3,6,7,3,2),
             b=c(7,8,4,2,3),
             c=c(3,4,5,6,1,2),
             d=c(1,5,3),
             e=c(5,3,1),
             f=c(3,1,5))
alignvecs <- rotate2front(vecs,first=1:8)
duplicated(alignvecs)

rotate2front(list(1:3,letters[3:1],letters[1:3],3:1))
rotate2front(c(4,2,3,1),first=1)

}
\author{
Sercan Kahveci
}
