% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{runs}
\alias{runs}
\title{Identify runs of identical values in a vector}
\usage{
runs(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
An integer vector of the same length as \code{x}, 
where each value identifies the run to which
the same-position value in \code{x} belongs.
}
\description{
Identify runs of identical values in a vector
}
\examples{
myvec <- sample(letters[1:3],size=20,replace=TRUE)
cbind(myvec,runs(myvec))

}
\author{
Sercan Kahveci
}
