% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_title_case.R
\name{safe_title_case}
\alias{safe_title_case}
\title{Safely convert titles to title case, preserving protected braces}
\usage{
safe_title_case(titles, component = NULL)
}
\arguments{
\item{titles}{character vector of BibTeX titles}

\item{component}{Either of 'title', 'booktitle' or 'journal'}
}
\value{
character vector of titles, wrapped in single braces, with protected content intact
}
\description{
Safely convert titles to title case, preserving protected braces
}
\keyword{internal}
