% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{scale.vector}
\alias{scale.vector}
\title{Standardize a vector}
\usage{
\method{scale}{vector}(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{Numeric vector to standardize.}

\item{center}{whether centering should occur (default \code{TRUE}).}

\item{scale}{Whether scaling should occur (default \code{TRUE}).}
}
\value{
Numeric vector with mean of 0 (if centered) and standard deviation of 1 (if scaled).
}
\description{
A method for scale() that takes and returns a vector.
}
\examples{
scale(1:10)

}
\author{
Sercan Kahveci
}
