% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{Semirandomize}
\alias{Semirandomize}
\title{Semi-randomize a sequence}
\usage{
Semirandomize(x, maxrep)
}
\arguments{
\item{x}{A vector to randomize.}

\item{maxrep}{Maximum permissible number of sequential occurrences of the same value.}
}
\value{
Vector \code{x} in semirandomized order, with no more than \code{maxrep}
sequential occurrences of the same value.
}
\description{
This algorithm first uses [base::sample()] to randomize the sequence and then 
repetitively eliminates excessively long subsequences of the same value 
by swapping individual values.
}
\details{
This gives an error if semirandomization is not possible, 
for example when one value in \code{x} is too numerous.
}
\examples{
Semirandomize(rep(0:1,each=96),maxrep=2)

Semirandomize(rep(0:2,c(1,4,1)),maxrep=2)

}
