% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_composite}
\alias{seq_composite}
\title{Generate and concatenate multiple integer sequences}
\usage{
seq_composite(from, to)
}
\arguments{
\item{from, to}{the starting and (maximal) end values of the sequences. 
Multiple can be given.}
}
\value{
An integer vector of multiple concatenated integer sequences.
}
\description{
This generates multiple integer sequences and concatenates them.
}
\examples{
seq_composite(from=c(1,7),to=c(3,8))

}
\author{
Sercan Kahveci
}
