% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{serr}
\alias{serr}
\title{Standard error}
\usage{
serr(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical. Should missing values be removed?}
}
\value{
The standard error of \code{x}. Returns \code{NA} for
vectors with non-removed missing values, 
as well as those with 1 or less non-\code{NA} values.
}
\description{
Compute the standard error of the mean.
}
\examples{
serr(rnorm(100))

}
\author{
Sercan Kahveci#'
}
