% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{setColNames}
\alias{setColNames}
\alias{setRowNames}
\title{Set column and row names of an object
These are convenience functions that return an object with its column or row names changed.
Use it in pipes.}
\usage{
setColNames(x, names)

setRowNames(x, names)
}
\arguments{
\item{x}{an object}

\item{names}{column or row names to be assigned to the object}
}
\description{
Set column and row names of an object
These are convenience functions that return an object with its column or row names changed.
Use it in pipes.
}
\examples{
setColNames(ToothGrowth,c("length","supplement","dosage"))
setRowNames(BOD,BOD$Time)
}
