% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{setunique}
\alias{setunique}
\alias{setduplicated}
\title{Set-based unique and duplicate detection}
\usage{
setunique(x, unique.only = FALSE)

setduplicated(x, unique.only = FALSE)
}
\arguments{
\item{x}{A list of vectors.}

\item{unique.only}{Should duplicated values within sets be ignored? Defaults to \code{FALSE}.}
}
\value{
For \code{setunique()}, a list of unique sets. 
For \code{setduplicated()}, a logical vector indicating whether
a set occurred previously.
}
\description{
These functions are like [base::unique()] and [base::duplicated()] except they only look at
whether two list elements contain the same values - the order does not matter.
}
\examples{
mysets <- list(a=c(1,2,3),b=2,c=c(3,2,1),d=c(1,3),e=c(1,1,2,3,3))
setunique(mysets)
setunique(mysets, unique.only=TRUE)

setduplicated(mysets)
setduplicated(mysets, unique.only=TRUE)

}
\author{
Sercan Kahveci
}
