% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skmisc-package.R
\docType{package}
\name{skmisc-package}
\alias{skmisc}
\alias{skmisc-package}
\title{skmisc: Miscellaneous Functions by Shafayet Khan}
\description{
Contains some function that makes certain workflows easier.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shafayetShafee/skmisc}
  \item \url{https://shafayetshafee.github.io/skmisc/}
  \item Report bugs at \url{https://github.com/shafayetShafee/skmisc/issues}
}

}
\author{
\strong{Maintainer}: Shafayet Khan Shafee \email{sshafee@isrt.ac.bd} (\href{https://orcid.org/0009-0002-8021-3788}{ORCID}) [copyright holder]

}
\keyword{internal}
