% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{smoothvect}
\alias{smoothvect}
\title{Smooth a numeric vector using a moving window algorithm}
\usage{
smoothvect(vect, width = 2, both.sides = T, alg = c("mean", "gauss"))
}
\arguments{
\item{vect}{}

\item{width}{Over how many values should the vector be averaged?}

\item{both.sides}{If TRUE (default), takes the mean of \code{width} values before and after the current index. If FALSE, only takes values ahead of the current index.}
}
\value{
Smoothed numeric vector
}
\description{
Smooth a numeric vector using a moving window algorithm
}
\examples{
temp<- smoothvect(beaver1$temp)
plot(temp,type="l")
}
