% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{sortmat}
\alias{sortmat}
\title{Sort a square matrix
This uses an iterative algorithm that swaps the rows and columns of a matrix 
to ensure the highest values are either in the middle or the bottom-right.}
\usage{
sortmat(mat, sorttype = c("diag", "center", "bottomright"))
}
\arguments{
\item{mat}{Square numeric matrix to be sorted.}

\item{sorttype}{Where the highest values should appear - 
"diag", "center", or "bottomright".}
}
\value{
A sorted square matrix.
}
\description{
Sort a square matrix
This uses an iterative algorithm that swaps the rows and columns of a matrix 
to ensure the highest values are either in the middle or the bottom-right.
}
\examples{

mat<-matrix(rnorm(144),ncol=12,nrow=12)
mat<-mat+t(mat)

newmat1<-sortmat(mat,"bottomright")

newmat2<-sortmat(mat,"diag")

newmat3<-sortmat(mat,"center")

}
\author{
Sercan Kahveci
}
