% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AATtools.R
\name{SpearmanBrown}
\alias{SpearmanBrown}
\title{Correct a correlation coefficient for being based on only a subset of the data.}
\usage{
SpearmanBrown(corr, ntests = 2)
}
\arguments{
\item{corr}{To-be-corrected correlation coefficient}

\item{ntests}{An integer indicating how many times larger the full test is, for which the corrected correlation coefficient is being computed.
When \code{ntests=2}, the formula will compute what the correlation coefficient would be if the test were twice as long.}
}
\value{
Spearman-Brown-corrected correlation coefficient. Values are bounded at zero.
}
\description{
Perform a Spearman-Brown correction on the provided correlation score.
}
