% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphlayout.R
\name{squarifyLayout}
\alias{squarifyLayout}
\title{Make a network layout more square-shaped}
\usage{
squarifyLayout(x, rate = 1)
}
\arguments{
\item{x}{A network layout matrix with 2 columns representing node x and y axis positions.}

\item{rate}{The rate by which to move nodes into the corners, between 0 and 1.}
}
\value{
A network layout matrix with 2 columns representing node x and y axis positions.
}
\description{
This transforms a circle-shaped network layout more into a square-shaped network layout. 
It does so by moving nodes in the corners more towards the corner.
}
