% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{std.errors}
\alias{std.errors}
\alias{serr}
\alias{serr.sd}
\alias{serr.var}
\alias{serr.skewness}
\alias{serr.kurtosis}
\title{Standard error estimators}
\usage{
serr(x, boot = 0, na.rm = FALSE)

serr.sd(x, boot = 0, na.rm = FALSE)

serr.var(x, boot = 0, na.rm = FALSE)

serr.skewness(x, na.rm = FALSE)

serr.kurtosis(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{boot}{If 0, SE will be determined algebraically. Otherwise,
this specifies the number of bootstraps used to compute the SE.}

\item{na.rm}{Logical. Should missing values be removed?}
}
\value{
The standard error of \code{x}. Returns \code{NA} for
vectors with non-removed missing values, 
as well as those with 1 or less non-\code{NA} values.
}
\description{
Compute the standard errors (SE) of various estimators, 
algebraically or through bootstrapping.
}
\section{Functions}{
\itemize{
\item \code{serr()}: SE of the mean

\item \code{serr.sd()}: SE of the standard deviation

\item \code{serr.var()}: SE of the variance

\item \code{serr.skewness()}: SE of the skewness

\item \code{serr.kurtosis()}: SE of the kurtosis

}}
\note{
The SEs of the skewness and kurtosis are computed for 
the second skewness and kurtosis formulas reported by 
Joannes and Gill (1998).

The algebraic estimators all assume a normal distribution. 
Results may be inaccurate if the data is distributed differently. 
In such occasions, bootstrapping should be preferred.
}
\examples{
serr(rnorm(100))

serr.sd(rnorm(100))

serr.var(rnorm(100))

}
\references{
Harding, B., Tremblay, C., & Cousineau, D. (2014). 
Standard errors: A review and evaluation of standard error estimators 
using Monte Carlo simulations. 
The Quantitative Methods for Psychology, 10(2), 107-123. 
doi: 10.20982/tqmp.10.2.p107

D. N. Joanes and C. A. Gill (1998). 
Comparing measures of sample skewness and kurtosis. 
The Statistician, 47, 183–189.
}
