% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stri_squish}
\alias{stri_squish}
\title{Trim unnecessary white-spaces from a string.}
\usage{
stri_squish(x)
}
\arguments{
\item{x}{string}
}
\value{
trimmed string
}
\description{
Trim unnecessary white-spaces from a string.
}
\keyword{internal}
