% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{strsplit.wrap}
\alias{strsplit.wrap}
\title{Create substrings with a maximal length by splitting at specific characters}
\usage{
strsplit.wrap(x, width = 2000, split = c("\\n", " ", ",", ""))
}
\arguments{
\item{x}{A character vector of length 1.}

\item{width}{The maximum character length to break the vector at.}

\item{split}{A vector of regular expressions to match a character to break the string at.
The function will try to break the string at the first value specified in this argument;
if that fails, it will move on to the second, then the third, etc.}
}
\value{
A character vector consisting of strings of length \code{width} or shorter,
and split at the characters specified in \code{split}.
}
\description{
This function splits a string into substrings of length \code{width} or shorter.
The splitting is done at the characters specified in \code{split}, in order of preference.
}
\details{
This combines the functionality of \code{\link[base:strwrap]{base::strwrap()}} and \code{\link[base:strsplit]{base::strsplit()}};
instead of a string wrapped with newlines, the result is multiple substrings.
}
\examples{
thanks <- paste(readLines(file.path(R.home("doc"), "THANKS")), collapse = "\n")
strsplit.wrap(thanks,width=80)

alphabet <- paste0(letters,collapse="")
strsplit.wrap(alphabet,width=3)
}
\author{
Sercan Kahveci
}
