% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{subdivide}
\alias{subdivide}
\title{Divide a vector or list}
\usage{
subdivide(x, divs, divlen)
}
\arguments{
\item{x}{the to-be-divided object.}

\item{divs, divlen}{The number of divisions and the preferred length of divisions. 
One and only one of \code{divs} and \code{divlen} must be given.}
}
\value{
A list consisting of \code{x}, divided in parts.
}
\description{
Divide a vector or list into parts of (preferably) equal length.
Either the length or the number of the parts can be set.
}
\examples{
subdivide(letters, divs=5)
subdivide(1:10, divlen=3)

}
\author{
Sercan Kahveci
}
