% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetworks.R
\name{symmetrizeNetwork}
\alias{symmetrizeNetwork}
\title{Make an adjacency matrix symmetric}
\usage{
symmetrizeNetwork(mat, rule = c("AND", "OR"))
}
\arguments{
\item{mat}{An adjacency matrix.}

\item{rule}{When two nodes are not bidirectionally connected, what should be done?
\code{"AND"} (default) deletes the edge, 
while \code{"OR"} keeps the halved edge weight.}
}
\value{
A symmetric adjacency matrix.
}
\description{
This makes a adjacency matrix symmetric by averaging each value with its transpose.
}
\details{
This follows the implementation of the IsingFit package.
}
\author{
Sercan Kahveci
}
