% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{tpars}
\alias{tpars}
\title{t-distribution fitter
This is a wrapper around [MASS::fitdistr()] specifically intended to
fit the t-distribution.}
\usage{
tpars(x, df = 30)
}
\arguments{
\item{x}{Vector of values to fit the t-distribution to}

\item{df}{Starting df value}
}
\value{
An object of class \code{"fitdistr"}.
}
\description{
t-distribution fitter
This is a wrapper around [MASS::fitdistr()] specifically intended to
fit the t-distribution.
}
\examples{
h<-rt(1000,df=3)*3+10
tpars(h)

}
