% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{TransformPlots}
\alias{TransformPlots}
\title{Data Transformation Plots}
\usage{
TransformPlots(x, type = c("none", "asinhrate", "sqrt", "bcp"))
}
\arguments{
\item{x}{A numeric vector.}

\item{type}{Names of transformations from the [trans()] to apply to \code{x}.}
}
\description{
Visualize how different transformations of the data 
will fit to a normal distribution.
}
\examples{
TransformPlots(mtcars$disp)

}
\author{
Sercan Kahveci
}
