% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceNetworks.R
\name{transitionNet}
\alias{transitionNet}
\title{Model a transition network from sequence data}
\usage{
transitionNet(
  data,
  predictors,
  decaytype = c("step", "flat", "accrual"),
  parameter,
  predtype = c("onset-to-onset", "recent-onset-to-onset", "recency-to-onset"),
  direction = 1,
  sampletype = "all",
  gammas = seq(0, 1, 0.1),
  alpha = 1,
  force.positive = FALSE,
  ncores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data.frame with 3 columns: sequence (character),
state (character), and date (numeric).}

\item{predictors}{Which states should be included in the network?}

\item{decaytype}{Once a state has occurred, how should its predictor look?
\itemize{
\item If "flat", the predictor is 1 after the state has occurred until it
no longer satisfies the condition set by the parameter argument.
\item If "step", the predictor is 1 immediately after the state has occurred,
and decays to 0 until it no longer satisfies the condition set by
the parameter argument.
\item If "accrual", the predictor is 1 after the state has occurred
until sequence end.
}}

\item{parameter}{This determines the largest temporal distance from
the occurrence of a state that is transformed into a nonzero value
as a predictor. The exact behavior of this argument is determined by
the \code{decaytype} argument. See Details for clarification.}

\item{predtype}{What type of predictor should be used? See Details.}

\item{direction}{In which direction should predictions be made?
1 means current states predict future states, -1 means current states predict
past states.}

\item{sampletype}{When the occurrence of a state is predicted, should every
sequence be included ("all") or only those that contain the state ("visitors")?}

\item{gammas}{Which EBIC gamma values should be used for model selection?
Higher values mean less edges are retained in the network.}

\item{alpha}{This is the elasticnet mixing parameter.
1 (default) means lasso regularization, 0 means ridge regularization, and
values in-between mix the two. See \code{\link[glmnet:glmnet]{glmnet::glmnet()}}.}

\item{force.positive}{Should model beta values be restricted to
the positive range?}

\item{ncores}{Number of models to run in parallel.
Be careful not to overload system memory.}

\item{verbose}{Produce verbose output in console?}
}
\description{
Model a transition network from sequence data
}
\section{Rationale}{
This is a method for analyzing sequences of states. The first occurrence
of each state is predicted with the recently preceding
first or most recent occurrences of other states
using lasso-regularized logistic regressions.
You set with the \code{preds} argument which states are predicted and
used as predictors. All other states are kept in the data,
but do not contribute to prediction nor are predicted.

The level of analysis (and the actual data analyzed) is the \emph{first}
occurrences of all states. That is, each row represents
the first occurrence of a state.
For each model, the dependent variable is coded
as 0 if the first occurrence of a state on the current row
is not the state being predicted, and 1 if it is that state;
prediction within a single sequence continues until
this state or the end of the sequence is reached.
Hence, if a state is reached before the end of the sequence,
all states after it are ignored in the prediction of that state,
since there can only be one first occurrence of a state in a sequence,
and therefore prediction of a first occurrence would not make sense
after the first occurrence had already occurred.
}

\section{Predictors}{
In a model predicting the first occurrence of state A,
all other selected states are used as predictors.
When \code{predtype} is \code{"onset-to-onset"},
the predictor for state B is 0 before and when that state first occurs, and
it is 1 for the first state occurring directly after.
Depending on \code{decaytype}, it tapers down (\code{"step"}) or
remains 1 (\code{"flat"}), until a number of states have passed
equal to \code{parameter}. Sensible values may thus include 1 to 10
(measured in first occurrences).

When \code{predtype} is \code{"recent-onset-to-onset"},
the predictor for state B is instead 1 or tapered down from 1 to 0
depending on the amount of time that has passed since
the first occurrence of state B; that amount of time is given by \code{parameter}.
Sensible values depend on the unit of time in the date column,
e.g. 24\*60\*60 for a day in seconds.

When \code{predtype} is \code{"recency-to-onset"}, the unit is again time,
but measured from the most recent occurrence of state B rather than the first.
Hence, if state B first occurred early on and was followed by many other states,
it would be 1 if it occurred again right before state A.
}

\author{
Sercan Kahveci
}
