% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faststats.R
\name{trimean}
\alias{trimean}
\title{Tukey's Trimean}
\usage{
trimean(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Whether to remove missing values. 
The function will error if \code{NA}s are present while \code{FALSE}.}
}
\value{
The trimean of \code{x}.
}
\description{
A robust mean estimator more efficient than the median, first proposed by Tukey.
}
\examples{
a <- rnorm(100)
a[1] <- 1000
trimean(a)
mean(a)

}
