% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{trypackages}
\alias{trypackages}
\title{Install packages if neccesary, then load them.}
\usage{
trypackages(...)
}
\arguments{
\item{...}{Unquoted names of packages to try loading, 
and if unable, install and load.}
}
\description{
Install packages if neccesary, then load them.
}
\examples{
trypackages(stats,utils,compiler)
}
