% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{unwrap.matrix}
\alias{unwrap.matrix}
\alias{rewrap.matrix}
\title{Convert between a matrix and a long-format data.frame}
\usage{
unwrap.matrix(x)

rewrap.matrix(x, na.value = NA)
}
\arguments{
\item{x}{In case of \code{unwrap.matrix()}, a matrix to unwrap; 
in case of \code{rewrap.matrix}, a data.frame with three columns,
respectively representing the row name, column name, and value.}

\item{na.value}{Which value to use in the matrix for elements 
not provided in \code{x}}
}
\value{
\code{unwrap.matrix()} returns a \code{data.frame} with three columns: 
\code{row} and \code{col} indicating the row and column names, and 
\code{value} indicating the respective value in the matrix. 
If no row or column names are available, 
the row or column number is used instead.

\code{rewrap.matrix()} returns a matrix.
}
\description{
Convert between a matrix and a long-format data.frame
}
\examples{
mymatrix <- matrix(1:40,ncol=8,nrow=5)
unwrap.matrix(mymatrix)
rewrap.matrix(unwrap.matrix(mymatrix))

carmatrix <- as.matrix(mtcars)
unwrap.matrix(carmatrix)
rewrap.matrix(unwrap.matrix(carmatrix))

}
\author{
Sercan Kahveci
}
