% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{vec.removeOLs}
\alias{vec.removeOLs}
\title{Remove outlying observations from a vector}
\usage{
vec.removeOLs(x, s = 3, make.na = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{Vector to remove outliers from}

\item{s}{If a value deviates more SDs from the mean than this value, it is marked as an outlier}

\item{make.na}{If \code{FALSE}, excludes the outliers. 
If \code{TRUE}, replaces them with \code{NA}.}

\item{verbose}{Report in console how many values have been removed?}
}
\value{
A vector with outliers removed or replaced with \code{NA}.
}
\description{
Remove outlying observations from a vector
}
\examples{
testvec <- c(1,3,5,7,9,11,13,15,17,19,100000)
vec.removeOLs(testvec)
vec.removeOLs(testvec,make.na=TRUE)

}
\seealso{
[removeOLs()]
}
\author{
Sercan Kahveci
}
