% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{vec.scale}
\alias{vec.scale}
\title{Scale a vector}
\usage{
vec.scale(x)
}
\arguments{
\item{x}{Numeric vector to standardize.}
}
\value{
Scaled numeric vector with mean of 0 and standard deviation of 1.
}
\description{
Like scale() but returns a vector and is faster.
}
\examples{
vec.scale(1:10)

}
\author{
Sercan Kahveci
}
