% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{vec2phrase}
\alias{vec2phrase}
\title{Convert a vector to an English list}
\usage{
vec2phrase(x)
}
\arguments{
\item{x}{A vector of values to convert into a string representing 
a grammatically correct English list.}
}
\value{
A string representing a grammatically correct English list.
}
\description{
Convert a vector to an English list
}
\examples{
vec2phrase(c("apples","oranges"))

vec2phrase(c("eggs"))

vec2phrase(c())

vec2phrase(c("cheese","milk","yoghurt","kefir"))

}
\author{
Sercan Kahveci
}
