% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{which.duplicate}
\alias{which.duplicate}
\title{Count duplicate values in a vector}
\usage{
which.duplicate(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector of the same length as \code{x}, where each element represents
the number of times the value in the same position in \code{x} has been repeated so far.
}
\description{
which.duplicate() determines for each element of a vector 
how many times its value has occurred so far.
}
\details{
It works similarly to [base::duplicated()] which only determines 
whether a value has occurred before and not how many times.
}
\examples{
which.duplicate(c(1,6,5,2,1,1,8,6,5))

}
\author{
Sercan Kahveci
}
