% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4tools.R
\name{AnovaTable}
\alias{AnovaTable}
\alias{print.AnovaTable}
\title{Compare multilevel models}
\usage{
AnovaTable(
  ...,
  fullmodel,
  models,
  serial = F,
  suppress = c("AIC", "deviance", "logLik")
)

\method{print}{AnovaTable}(x)
}
\arguments{
\item{...}{Model objects to be compared}

\item{fullmodel}{A model to which all other models are to be compared; only use if \code{...} is not specified.}

\item{models}{Models to compare to \code{fullmodel}. Only use if \code{...} is not specified.}

\item{serial}{If TRUE, models are compared serially; if false, all models will be compared to the first.}

\item{suppress}{Character vector of column names to suppress in printed output.}
}
\value{
A data.frame containing model fit metrics such as AIC, BIC, marginal R-squared (the effect size of fixed effects only),
conditional R-squared (the effect size of all model terms), loglikelihood, deviance, and a likelihood ratio test.
}
\description{
Compare multilevel models
}
\examples{

}
