% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{AutocorPlot}
\alias{AutocorPlot}
\title{Multi-series Autocorrelation Plotting}
\usage{
AutocorPlot(x, index, lag.max = 64, plot = TRUE)
}
\arguments{
\item{x}{Variable to compute the autocorrelation of.}

\item{index}{Grouping ID variable of the same length as \code{X}.}

\item{lag.max}{The maximum lag at which to compute autocorrelation.}

\item{plot}{Should a plot be made? Default is \code{TRUE}.}
}
\description{
Multi-series Autocorrelation Plotting
}
\examples{
mycors<-AutocorPlot(x=ToothGrowth$len,index=ToothGrowth$supp,lag.max=10)

}
\author{
Sercan Kahveci
}
