% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_title_to_title_case.R
\name{bib_title_to_title_case}
\alias{bib_title_to_title_case}
\title{Convert the title, booktitle, journal name of each bibentry of a bib file to title case.}
\usage{
bib_title_to_title_case(...)
}
\value{
The return value of \code{\link[=bib_title_case]{bib_title_case()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{bib_title_to_title_case()} has been deprecated in \strong{v0.3.0} and will be
removed in a future release.
Please use \code{\link[=bib_title_case]{bib_title_case()}} instead.

This function remains as a compatibility wrapper and simply forwards all
arguments to \code{bib_title_case()}.
}
\details{
This function takes any arguments (\code{...}) and passes them directly to
\code{\link[=bib_title_case]{bib_title_case()}}. See that function's documentation for full argument
details.
}
\seealso{
\code{\link[=bib_title_case]{bib_title_case()}}
}
\keyword{internal}
