% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_title_to_title_case.R
\name{bib_title_to_title_case}
\alias{bib_title_to_title_case}
\title{Convert the title of each bibentry of a bib file to title case.}
\usage{
bib_title_to_title_case(bib_file_path)
}
\arguments{
\item{bib_file_path}{character, path to a \code{.bib} file.}
}
\value{
\code{bib_file_path} as a character string, invisibly.
}
\description{
Convert the title of each bibentry of a bib file to title case.
}
