% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faststats.R
\name{bonferroniHolm}
\alias{bonferroniHolm}
\title{Bonferroni-Holm significance evaluation
Evaluate the significance of a vector of p-values according to
the Bonferroni-Holm method}
\usage{
bonferroniHolm(x, alpha = 0.05)
}
\arguments{
\item{x}{A vector of p-values.}

\item{alpha}{An experiment-wise alpha-level to test against.}
}
\value{
A vector indicating whether values in \code{x} were significant
according to the Bonferroni-Holm method.
}
\description{
Bonferroni-Holm significance evaluation
Evaluate the significance of a vector of p-values according to
the Bonferroni-Holm method
}
\examples{
bonferroniHolm(c(1,.05,.001,.002,.01))

}
\author{
Sercan Kahveci
}
