% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funs.R
\name{check_numeric_scalar}
\alias{check_numeric_scalar}
\title{Check whether an object is a valid numeric scalar}
\usage{
check_numeric_scalar(x)
}
\arguments{
\item{x}{An R object to check.}
}
\value{
Invisibly returns \code{TRUE} on check pass. Throws a error of class
\code{\link[rlang:rlang_error]{rlang::rlang_error}} on check fail.
}
\description{
This function checks whether an object \code{\link[=is_numeric_scalar]{is_numeric_scalar()}}. If the check
fails, a formatted error is raised using \code{\link[cli:cli_abort]{cli::cli_abort()}}.
}
\examples{
check_numeric_scalar(5)       # passes
check_numeric_scalar(0.1)     # passes
\dontrun{
check_numeric_scalar(NA)      # throws error
check_numeric_scalar(NULL)    # throws error
check_numeric_scalar(c(1,2))  # throws error
check_numeric_scalar("5")     # throws error
}

}
