% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{clamp}
\alias{clamp}
\title{clamp}
\usage{
clamp(val, minval, maxval)
}
\arguments{
\item{val}{The vector/matrix to clamp}

\item{minval}{Minimum value; all lower values are clamped to this value}

\item{maxval}{Maximum value; all higher values are clamped to this value}
}
\value{
Clamped vector.
}
\description{
clamp
}
\examples{
clamp(0:10,2,8)
}
