% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_article_template.R
\name{create_article_template}
\alias{create_article_template}
\title{Create a structured article template for a specified journal}
\usage{
create_article_template(journal_name)
}
\arguments{
\item{journal_name}{A character string naming the journal for which the
article template will be created. It must be one of the journals listed
by \code{rticles::journals()}.}
}
\value{
This function returns no value. It creates the directory
structure for the article template, including the required \code{Rmd}
files for sections and the abstract.
}
\description{
Generate an article template for a given journal using
\code{rmarkdown::draft} from the \code{rticles} package. The function
checks whether the journal is supported, creates needed directories and
files, and formats the \code{<journal_name>_article.Rmd} file with
section chunks and an abstract.
}
\examples{
\dontrun{
create_article_template("arxiv")
}
}
