% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{df.init}
\alias{df.init}
\title{Initiate an empty data frame}
\usage{
df.init(x)
}
\arguments{
\item{x}{A character vector of column names.}
}
\value{
A data.frame with 0 rows.
}
\description{
Initiate an empty data frame
}
\examples{
test <- df.init(c("A","B","C"))

}
