% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skMisc.R
\name{df.init}
\alias{df.init}
\title{Initiate an empty data frame}
\usage{
df.init(namelist)
}
\arguments{
\item{namelist}{A character vector of column names.}
}
\value{
A data.frame with 0 rows.
}
\description{
Initiate an empty data frame
}
