% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{distdiff}
\alias{distdiff}
\title{Compare distributions}
\usage{
distdiff(x, y, op = c(">", "<", ">=", "<=", "sign"), paired = F)
}
\arguments{
\item{x, y}{Vectors to compare.}

\item{op}{The comparison operation to perform.}

\item{paired}{Are the values from \code{x} and \code{y} paired?
If \code{TRUE}, each element in \code{x} is compared only to the element in the same position in \code{y};
otherwise, every element in \code{x} is compared to every element in \code{y}.}
}
\value{
The requested probability estimate as a single value.
}
\description{
These are fast routines to compute the probability that a random draw from x
exceeds a random draw from y (\code{op=">"}) or
the difference between the probability of x>y minus the difference of the probability of x<y (\code{op="sign"}).
}
\examples{
a<-rnorm(1000)
b<-rnorm(1000)+1
distdiff(a,b,">")
distdiff(a,b,"<")
distdiff(a,b,"sign")

}
\author{
Sercan Kahveci
}
