% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_char_scalar}
\alias{is_char_scalar}
\title{Check if x is a single, non-NA character}
\usage{
is_char_scalar(x)
}
\arguments{
\item{x}{any object}
}
\value{
TRUE if x is a single, non-NA character, FALSE otherwise
}
\description{
Check if x is a single, non-NA character
}
\keyword{internal}
