% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_funs.R
\name{is_invalid}
\alias{is_invalid}
\title{Test whether an object is considered invalid}
\usage{
is_invalid(x)
}
\arguments{
\item{x}{An R object to check.}
}
\value{
A logical scalar: \code{TRUE} if the object is invalid, \code{FALSE} otherwise.
}
\description{
This function tests whether an R object should be considered invalid. An
object is invalid if it meets any of the following conditions:
\itemize{
\item it is \code{\link[base:missing]{base::missing()}},
\item it is \code{NULL},
\item it has zero length,
\item it inherits from class \code{"try-error"},
\item it inherits from class \code{"simpleError"},
\item it is a list whose elements are \emph{all} invalid (checked recursively), or
\item it is an atomic vector whose elements are \emph{all} \code{NA}.
}
}
\details{
Lists are validated recursively: a list is invalid only if \emph{every} element
in that list is invalid. Atomic vectors are considered invalid only when
\emph{all} values are \code{NA}.
}
\examples{
is_invalid(NULL)                 # TRUE
is_invalid(character(0))         # TRUE
is_invalid(c(NA, NA))            # TRUE
is_invalid(c(1, NA))             # FALSE

is_invalid(list(NA, NULL))       # TRUE
is_invalid(list(NA, 5))          # FALSE

# try-error example
err1 <- try(stop("err"), silent = TRUE)
is_invalid(err1)                 # TRUE

# simpleError example
err2 <- simpleError("oops")
is_invalid(err2)                 # TRUE

is_invalid(5)                    # FALSE
is_invalid("txt")                # FALSE

}
