% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{LevenshteinDistance}
\alias{LevenshteinDistance}
\title{Levenshtein distance}
\usage{
LevenshteinDistance(x, y)
}
\arguments{
\item{x, y}{Strings to be compared.}
}
\value{
The Levenshtein distance between the two strings.
}
\description{
Counts the number of single character deletions, insertions, and substitutions 
that need to be performed to turn the source string into the target string.
}
\examples{
LevenshteinDistance("sitting","kitten")

LevenshteinDistance(c("cheese","fish"),c("child","flan"))

}
\author{
Sercan Kahveci
}
