% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrangler.R
\name{match.merge}
\alias{match.merge}
\title{Match and merge two data.frames by ID and date}
\usage{
match.merge(x, y, idvar, datevar, keep.unmatched = c("all", "x", "y", "none"))
}
\arguments{
\item{x, y}{The \code{data.frame}s to merge.}

\item{idvar}{Name of the variable holding participant/session IDs in both x and y.}

\item{datevar}{Name of the variable holding dates in both x and y.}

\item{keep.unmatched}{When no match is found for a row, should it be kept or removed?}
}
\value{
A merged \code{data.frame} where rows of \code{x}
are merged with rows of \code{y} that match by session id and where
the date of each row from \code{x} directly precedes the 
date of the matched row from \code{y}.
}
\description{
the rows from \code{y} are matched with rows from \code{x} that 
share the same ID and directly precede them in date.
}
\details{
This is useful for matching participant data from different datasets,
where one dataset is always collected after another. It is assumed here that
\code{x} is always administered before \code{y}.

This matches using [match.pps()] and merges using [dplyr::full_join()].
}
\examples{
x <- data.frame(id=c(1,2,2,3,3,4),
                date=as.POSIXct(c(1,1,3,1,3,1)),
               info=letters[1:6])
y <- data.frame(id=c(1,2,3,3,4),
               date=as.POSIXct(c(2,2,4,2,0)),
                data=LETTERS[1:5])
match.merge(x=x,y=y,idvar="id",datevar="date",keep.unmatched="all")
match.merge(x=x,y=y,idvar="id",datevar="date",keep.unmatched="none")

}
