% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{modular.mean}
\alias{modular.mean}
\title{Title}
\usage{
modular.mean(x, mod, check = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector of modular quantities to be averaged.}

\item{mod}{The modulus of \code{x}, i.e. where the numbers "loop back" to zero. 
On a clock this is 12, in degrees it is 360, etc.}

\item{check}{Should a check be performed to detect whether there is no identifiable mean?
This should be set to \code{FALSE} when such a situation is not expected in the data, since
this check can be computationally costly.}

\item{na.rm}{Should \code{NA} values be removed? Defaults to \code{TRUE}.}
}
\description{
Title
}
\details{
This function finds the value that has the smallest squared distance 
to all values in \code{x}.
}
\examples{


# Situations where no modular mean exists (points are evenly divided around the circle)
modular.mean(c(0,3,6,9),mod=12)
modular.mean(c(0,1,3,4,6,7,9,10),mod=12)

}
