% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{mostcommon}
\alias{mostcommon}
\title{Return most common element(s) of vector}
\usage{
mostcommon(x, n = 1)
}
\arguments{
\item{x}{Vector.}

\item{n}{Number of most common elements to return.}
}
\value{
A vector of the most common elements of \code{x}, with length \code{n}.
}
\description{
Return most common element(s) of vector
}
\examples{
mostcommon(c(1,2,2,3,7,6,4,4,4))
mostcommon(c(1,2,2,3,7,6,4,4,4), n=2)

}
