% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{setunique}
\alias{setunique}
\alias{setduplicated}
\title{Set-based unique and duplicate detection}
\usage{
setunique(x)

setduplicated(x)
}
\arguments{
\item{x}{A list of vectors}
}
\value{
For \code{setunique()}, a list of unique sets. 
For \code{setduplicated()}, a logical vector indicating whether
a set occurred previously.
}
\description{
These functions are like [base::unique()] and [base::duplicated()] except they only look at
whether two list elements contain the same values - the order does not matter.
}
\examples{
mysets <- list(a=1:3,b=2,c=3:1,d=c(1,3))
setunique(mysets)
setduplicated(mysets)

}
\author{
Sercan Kahveci
}
