% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skMisc.R
\docType{package}
\name{skMisc}
\alias{skMisc-package}
\alias{skMisc}
\title{skMisc: Miscellaneous convenience functions}
\description{
Functionality provided

Some of the functions in this package are statistical:
\itemize{
\item Common outlier exclusion methods and statistical analyses with concise print functions
\item Functions for manipulating lme4 formulas
\item Methods for obtaining "holdout statistics"
}

Some of the functions are geared towards datawrangling:
\itemize{
\item Functions for merging, subsetting, and transforming data.frames and matrices
\item Specific date and string manipulations unavailable in common packages
\item Other small convenience functions that make life easier
}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/Spiritspeak/skMisc/issues}
}

}
\author{
\strong{Maintainer}: Sercan Kahveci \email{sercan+kahveci@sbg.ac.at}

}
