% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{splitColumn}
\alias{splitColumn}
\title{Split a character column into multiple values}
\usage{
splitColumn(x, sep = ";")
}
\arguments{
\item{x}{a character vector to split into columns}

\item{sep}{a caracter separating the different values}
}
\value{
a \code{data.frame} of boolean values, with rows representing the unpacked
vector entries and columns indicating whether the specific value
}
\description{
Split a character column into multiple values
}
\examples{
unsplit<-c("flour;salt;baking soda;steak;sugar;water;sauce;vinegar",
"flour;sauce;mustard;salt;pepper;vinegar;baking soda;water;tomatoes;onion;steak")
splitColumn(unsplit)
}
