% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{theme_pecher}
\alias{theme_pecher}
\title{Pecher theme for ggplot
Based on the plot design style of prof. Diane Pecher.}
\usage{
theme_pecher()
}
\value{

}
\description{
Pecher theme for ggplot
Based on the plot design style of prof. Diane Pecher.
}
