% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{trans}
\alias{trans}
\title{Vector transformations}
\usage{
trans(
  x,
  type = c("none", "log", "log1p", "sqrt", "inv", "atan", "asinh", "logshift",
    "invshift", "asinhrate", "bcp", "bcnp", "yjp"),
  par = NULL
)
}
\arguments{
\item{x}{Vector to be transformed.}

\item{type}{Can be "none" (no transform), "log" (\code{log(x)}), "log1p" (\code{log(x+1)}),
"sqrt" (\code{sqrt(x)}), "inv" (\code{1/x}), "atan" (\code{atan(x)}),
"asinh" (\code{asinh(x-mean(x)) + mean(x)}),
"logshift" (\code{log(x+shift)-log(shift)} if \code{shift>0}, else \code{log(x+shift)}),
"invshift" (\code{(1+shift)/(x+1+shift)}),
"asinhrate" (\code{asinh((x-mean(x)) * rate) / rate + mean(x)}),
"bcp" (Box-Cox transform from \code{\link[car:bcPower]{car::bcPower()}}),
"bcnp" (Shifted Box-Cox transform from \code{\link[car:bcPower]{car::bcnPower()}}),
or "yj" (Yeo-Johnson transformation from \code{\link[car:bcPower]{car::yjPower()}}). Non-positive values in \code{x} are
only allowed in "none", "asinh", "asinhrate", "bcnp" and "yj".}

\item{par}{A parameter that determines the shape of the transformation. If no parameter is given, then the function finds
\itemize{
\item for "logshift" and "invshift", the parameter that minimizes the skew
\item for "asinhrate", the parameter that minimizes the excess kurtosis
\item for "bcp", "bcnp", and "yj", the parameter that gives the most normally distributed values, using [\link[car:powerTransform]{car::powerTransform}]
For \code{"logshift"} and \code{"invshift"}, this is the shift;
For \code{"asinhrate"}, this is the rate, with higher values representing a stronger reduction of kurtosis;
for \code{"bcp"} and \code{"yj"}, this is the lambda parameter;
for \code{"bcnp"}, this is two values: the lambda and gamma parameters.
For all other transformations, \code{par} is ignored.
}}
}
\value{
A transformed vector.
Attribute \code{"transform.method"} indicates the used method, and
\code{"transform.par"} indicates the transformation parameter(s).
}
\description{
Normalizing transformations for skewed and leptokurtic variables
}
\details{
\itemize{
\item For skewed distributions, consider "log", "sqrt", "inv", "atan", "logshift", "invshift", "bcp", "bcnp", or "yjp".
\item For distributions with excess kurtosis, consider "asinh" or "asinhrate".
}
}
\examples{
# Normality comparison
set.seed(1)
u<-rgamma(1000,2)
logtrans<-trans(u,"log")
shapiro.test(logtrans) # Not normal
logshifttrans<-trans(u,"logshift")
shapiro.test(logshifttrans) # Not normal but less bad
bctrans<-trans(u,"bcp")
shapiro.test(bctrans) # Normal

# Shifted distribution
u<-rgamma(1000,2)+2
logshifttrans<-trans(u,"logshift")
shapiro.test(logshifttrans) # Not normal 
invshifttrans<-trans(u,"invshift")
shapiro.test(invshifttrans) # Not normal


u<-rt(1000,2)+10 
yjtrans<-trans(u,"yjp")
shapiro.test(yjtrans) # Not normal; still leptokurtic
asinhtrans<-trans(u,"asinh")
shapiro.test(asinhtrans) # Normal!

}
\seealso{
\code{\link[=removeOLs]{removeOLs()}}, \code{\link[=vec.removeOLs]{vec.removeOLs()}}
}
\author{
Sercan Kahveci
}
