% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{TransformPlots}
\alias{TransformPlots}
\title{Data Transformation Plots}
\usage{
TransformPlots(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Visualize how different transformations of the data 
will fit to a normal distribution.
}
\examples{
TransformPlots(mtcars$disp)

}
\author{
Sercan Kahveci
}
