% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skUtils.R
\name{unwrap.matrix}
\alias{unwrap.matrix}
\title{Turn a matrix into a long-format data.frame}
\usage{
unwrap.matrix(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
a \code{data.frame} with three columns: \code{row} and \code{col} indicating the
row and column names, and \code{value} indicating the respective value in the matrix. 
If no row or column names are available, the row or column number is used instead.
}
\description{
Turn a matrix into a long-format data.frame
}
\examples{
mymatrix<-matrix(1:80,ncol=8,nrow=10)
unwrap.matrix(mymatrix)

carmatrix<-as.matrix(mtcars)
unwrap.matrix(carmatrix)

}
